+++ CHANGELOG.md
@@ -14,6 +14,11 @@ Latest
 * Added `String.__mul__` to support easy repetition of strings, e.g.,
   `"asdf" * 3 ==> "asdfasdfasdf"`.
 
+### Packaging
+
+* The `quark` standard library package is now shaded in Java so that its
+  dependencies don't conflict with user code's dependencies.
+
 1.0.332
 ------